Restructure PebbloRetrievalQA code
